+++ common/src/main/java/smartin/miapi/editor/syntax/CodecMaterialEditorInterface.java
@@ -48,7 +48,8 @@ public class CodecMaterialEditorInterface implements EditorInterface {
                             Optional.empty(),
                             Optional.empty(),
                             new ArrayList<>(),
-                            Either.left(false)
+                            Either.left(false),
+                            List.of()
                     )).getOrThrow()).result();
 
             if (result.isEmpty()) {